+++ BoostHTMLServer3/HeaderManager.h
@@ -14,14 +14,14 @@ namespace websocket
 	class HeaderManager
 	{
 	public:
-		static const size_t HEADER_OUT_SIZE = 2;
-		static const size_t HEADER_IN_SIZE = 2;
+		static const unsigned int HEADER_OUT_SIZE = 2;
+		static const unsigned int HEADER_IN_SIZE = 2;
 
 		HeaderManager();
 
 		virtual void encryptHeader(const OPacket* pack);
 
-		virtual IPacket* const decryptHeader(char* header, size_t size, bool& serverRead, dataframe* dataFrame, IDType cID);
+		virtual IPacket* const decryptHeader(char* header, unsigned int size, bool& serverRead, dataframe* dataFrame, IDType cID);
 
 		virtual ~HeaderManager();
 
@@ -30,9 +30,9 @@ namespace websocket
 
 		virtual void encryptHeaderToBigEndian(const OPacket* pack);
 
-		virtual IPacket* const decryptHeaderAsBigEndian(char* header, size_t size, bool& serverRead, dataframe* dataFrame, IDType cID);
+		virtual IPacket* const decryptHeaderAsBigEndian(char* header, unsigned int size, bool& serverRead, dataframe* dataFrame, IDType cID);
 
-		virtual IPacket* const decryptHeaderFromBigEndian(char* header, size_t size, bool& serverRead, dataframe* dataFrame, IDType cID);
+		virtual IPacket* const decryptHeaderFromBigEndian(char* header, unsigned int size, bool& serverRead, dataframe* dataFrame, IDType cID);
 
 		bool bEndian;
 	};
